refactor: use class in repository
